Make sensor info and platform events arrays dynamic
Driver init is a bit messier as a result.

container_of(), which we were using in some debugfs show
functions to get a pointer back to the driver state, does not
work if a struct array field is dynamically allocated.

Store the needed pointer in each sensor info struct in
addition to the WMI object instance number we already keep
and rework the affected debugfs show functions.