Fix failure to load on EliteDesk 800 G6
The EliteDesk 800 G6 stores a raw WMI string within the ACPI object
in its BIOS corresponding to one instance of HPBIOS_PlatformEvents.Name.
This is evidently a valid way of representing a WMI data item as far
as the Microsoft ACPI-WMI mapper is concerned, but is preventing the
driver from loading.

This seems quite rare, but add support for such strings. Treating
this as a quirk pretty much means adding that support anyway.

Also clean up an oversight in update_numeric_sensor_from_wobj() in
which the result of hp_wmi_strdup() was being used without error
checking.

Tracks c9ba592580 ("hwmon: (hp-wmi-sensors) Fix failure to load
on EliteDesk 800 G6") from mainline.

Link: https://lore.kernel.org/r/20231123054918.157098-1-james@equiv.tech