Check if WMI event data exists
The BIOS can choose to return no event data in response to a
WMI event, so the ACPI object passed to the WMI notify handler
can be NULL.

Check for such a situation and ignore the event in such a case.

Tracks a54da9df75 ("hwmon: (hp-wmi-sensors) Check if WMI event
data exists") from mainline.

Link: https://lore.kernel.org/all/20240901031055.3030-2-W_Armin@gmx.de/